requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.57.0
django-filter==23.3
django-rest-auth==0.9.5
djangorestframework==3.14.0
idna==3.4
Markdown==3.5
pycparser==2.21
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
